app/account/requirements.txt
flask
websockets
requests
python-binance
python-dotenv
